Move remove-undefined-objects to a new function
We can add more edge case tests easier this way
