+++ packages/api-explorer/src/lib/oas-to-har.js
@@ -3,6 +3,7 @@ const querystring = require('querystring');
 const extensions = require('@readme/oas-extensions');
 const getSchema = require('./get-schema');
 const configureSecurity = require('./configure-security');
+const removeUndefinedObjects = require('./remove-undefined-objects');
 
 // const format = {
 //   value: v => `__START_VALUE__${v}__END__`,
@@ -72,42 +73,6 @@ function getResponseContentType(content) {
 
 function isPrimitive(val) {
   return typeof val === 'string' || typeof val === 'number' || typeof val === 'boolean';
-}
-
-function isEmptyObject(obj) {
-  // Then remove all empty objects from the top level object
-  return typeof obj === 'object' && Object.keys(obj).length === 0;
-}
-
-// Modified from here: https://stackoverflow.com/a/43781499
-function stripEmptyObjects(obj) {
-  Object.keys(obj).forEach(key => {
-    const value = obj[key];
-    if (typeof value === 'object') {
-      // Recurse, strip out empty objects from children
-      stripEmptyObjects(value);
-      // Then remove all empty objects from the top level object
-      if (isEmptyObject(value)) {
-        delete obj[key];
-      }
-    }
-  });
-}
-
-function removeUndefinedObjects(obj) {
-  // JSON.stringify removes undefined values
-  const withoutUndefined = JSON.parse(JSON.stringify(obj));
-
-  // Then we recursively remove all empty objects
-  stripEmptyObjects(withoutUndefined);
-
-  // If the only thing that's leftover is an empty object
-  // then return nothing so we don't end up with default
-  // code samples with:
-  // --data '{}'
-  if (isEmptyObject(withoutUndefined)) return undefined;
-
-  return withoutUndefined;
 }
 
 module.exports = (
